requirements.txt
python-telegram-bot==20.4
openai==0.27.8
isort==5.12.0
black==23.7.0